refactor : 헥사고날 아키텍쳐로 리팩토링 한다
