+++ src/main/java/calculator/adapters/input/InputReaderImpl.java
@@ -1,5 +1,6 @@
-package calculator.view.input;
+package calculator.adapters.input;
 
+import calculator.application.ports.input.InputReader;
 import camp.nextstep.edu.missionutils.Console;
 
 public class InputReaderImpl implements InputReader {
