--- src/main/java/calculator/presentation/CalculationUseCase.java
@@ -1,6 +1,0 @@
-package calculator.presentation;
-
-public interface CalculationUseCase {
-
-    String handleCalculation();
-}
